+++ android/src/main/java/com/salesforce/marketingcloud/reactnative/RNSFMCSdkModule.java
@@ -26,9 +26,7 @@
 package com.salesforce.marketingcloud.reactnative;
 
 import android.util.Log;
-
 import androidx.annotation.NonNull;
-
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
@@ -46,18 +44,23 @@ import com.salesforce.marketingcloud.sfmcsdk.components.events.Event;
 import com.salesforce.marketingcloud.sfmcsdk.components.identity.Identity;
 import com.salesforce.marketingcloud.sfmcsdk.components.logging.LogLevel;
 import com.salesforce.marketingcloud.sfmcsdk.components.logging.LogListener;
+import com.salesforce.marketingcloud.sfmcsdk.modules.ModuleInterface;
 import com.salesforce.marketingcloud.sfmcsdk.modules.push.PushModuleInterface;
 import com.salesforce.marketingcloud.sfmcsdk.modules.push.PushModuleReadyListener;
-
 import java.util.Map;
 import java.util.Set;
-
 import javax.annotation.Nonnull;
 
 @SuppressWarnings({"unused", "WeakerAccess"})
 public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
     public RNSFMCSdkModule(ReactApplicationContext reactContext) {
         super(reactContext);
+        handlePushAction(new MCPushAction() {
+            @Override
+            void execute(PushModuleInterface sdk) {
+                sdk.getRegistrationManager().edit().addTag("React").commit();
+            }
+        });
     }
 
     @Override
@@ -242,6 +245,54 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
         SFMCSdk.track(sfmcEvent);
     }
 
+    @ReactMethod
+    public void isAnalyticsEnabled(Promise promise) {
+        handlePushAction(new MCPushAction() {
+            @Override
+            void execute(PushModuleInterface sdk) {
+                promise.resolve(sdk.getAnalyticsManager().areAnalyticsEnabled());
+            }
+        });
+    }
+
+    @ReactMethod
+    public void setAnalyticsEnabled(final Boolean enable) {
+        handlePushAction(new MCPushAction() {
+            @Override
+            void execute(PushModuleInterface sdk) {
+                if (enable) {
+                    sdk.getAnalyticsManager().enableAnalytics();
+                } else {
+                    sdk.getAnalyticsManager().disableAnalytics();
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void isPiAnalyticsEnabled(Promise promise) {
+        handlePushAction(new MCPushAction() {
+            @Override
+            void execute(PushModuleInterface sdk) {
+                promise.resolve(sdk.getAnalyticsManager().arePiAnalyticsEnabled());
+            }
+        });
+    }
+
+    @ReactMethod
+    public void setPiAnalyticsEnabled(final Boolean enable) {
+        handlePushAction(new MCPushAction() {
+            @Override
+            void execute(PushModuleInterface sdk) {
+                if (enable) {
+                    sdk.getAnalyticsManager().enablePiAnalytics();
+                } else {
+                    sdk.getAnalyticsManager().disablePiAnalytics();
+                }
+            }
+        });
+    }
+
     private void handleAction(final SFMCAction action) {
         SFMCSdk.requestSdk(new SFMCSdkReadyListener() {
             @Override
@@ -259,6 +310,11 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
                     @Override
                     public void ready(@NonNull PushModuleInterface pushModuleInterface) {
                         action.execute(pushModuleInterface);
+                    }
+
+                    @Override
+                    public void ready(@NonNull ModuleInterface moduleInterface) {
+                        this.ready((PushModuleInterface) moduleInterface);
                     }
                 });
             }
@@ -285,8 +341,7 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
     abstract class SFMCAction {
         abstract void execute(SFMCSdk sdk);
 
-        void err() {
-        }
+        void err() {}
     }
 
     abstract class SFMCPromiseAction extends SFMCAction {
@@ -304,7 +359,7 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
         @Override
         void err() {
             promise.reject("SFMCSDK-INIT",
-                    "The MarketingCloudSdk#init method must be called in the Application's onCreate.");
+                "The MarketingCloudSdk#init method must be called in the Application's onCreate.");
         }
 
         abstract void execute(SFMCSdk sdk, @NonNull Promise promise);
@@ -313,8 +368,7 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
     abstract class MCPushAction {
         abstract void execute(PushModuleInterface pushSdk);
 
-        void err() {
-        }
+        void err() {}
     }
 
     abstract class MCPushPromiseAction extends MCPushAction {
@@ -332,7 +386,7 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
         @Override
         void err() {
             promise.reject("SFMCSDK-INIT",
-                    "The MarketingCloudSdk#init method must be called in the Application's onCreate.");
+                "The MarketingCloudSdk#init method must be called in the Application's onCreate.");
         }
 
         abstract void execute(PushModuleInterface sdk, @NonNull Promise promise);
@@ -341,8 +395,7 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
     abstract class SFMCIdentityAction {
         abstract void execute(Identity identity);
 
-        void err() {
-        }
+        void err() {}
     }
 
     abstract class SFMCIdentityPromiseAction extends SFMCIdentityAction {
@@ -360,10 +413,9 @@ public class RNSFMCSdkModule extends ReactContextBaseJavaModule {
         @Override
         void err() {
             promise.reject("SFMCSDK-INIT",
-                    "The SFMCSdk#configure method must be called in the Application's onCreate.");
+                "The SFMCSdk#configure method must be called in the Application's onCreate.");
         }
 
         abstract void execute(Identity sdk, @NonNull Promise promise);
     }
-
 }